# openapi.yml
{"openapi":"3.0.1","info":{"title":"Bank app","description":"Тестовое задание"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Аутентификация","description":"Методы для регистрации и получения токенов доступа"},{"name":"Пользователи","description":"Позволяет получить информацию о пользователях"},{"name":"Банковские операции","description":"Операции перевода денег между счетами"}],"paths":{"/api/transfer":{"post":{"tags":["Банковские операции"],"summary":"Перевод денег","description":"Перевод денег со счета текущего аутентифицированного клиента на счет другого клиента","operationId":"performTransfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResult"}}}}}}},"/api/signup":{"post":{"tags":["Аутентификация"],"summary":"Регистрация пользователя","description":"Позволяет создать нового пользователя","operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResult"}}}}}}},"/api/refresh":{"post":{"tags":["Аутентификация"],"summary":"Обновить токены","description":"Позволяет обменять refresh token на новые access token и refresh token","operationId":"updateTokens","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTokenRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokensPair"}}}}}}},"/api/login":{"post":{"tags":["Аутентификация"],"summary":"Вход в аккаунт","description":"Позволяет обменять учетные данные на токены доступа","operationId":"loginClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokensPair"}}}}}}},"/api/client":{"get":{"tags":["Пользователи"],"summary":"Информация о текущем пользователе","description":"Позволяет получить инофрмацию о текущем пользователе","operationId":"getCurrentClient","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientSearchResponse"}}}}}},"patch":{"tags":["Пользователи"],"summary":"Изменение данных текущего пользователя","description":"Позволяет изменить эл. почты и номера телефонов текущего пользователя","operationId":"updateClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InformationUpdateRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDescription"}}}}}}},"/api/client/search":{"get":{"tags":["Пользователи"],"summary":"Поиск пользователей","description":"Позволяет выполнить поиск пользователей по критериям","operationId":"searchClient","parameters":[{"name":"phone","in":"query","description":"Номер телефона","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","description":"Электронная почта","required":false,"schema":{"type":"string"}},{"name":"fullName","in":"query","description":"ФИО","required":false,"schema":{"type":"string"}},{"name":"birthDate","in":"query","description":"Дата рождения","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Номер страницы","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","description":"Рамер страницы","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sortBy","in":"query","description":"Поле для сортировки","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceErrorMessage"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientSearchResponse"}}}}}}}},"components":{"schemas":{"ServiceErrorMessage":{"type":"object","properties":{"error":{"type":"string"}}},"TransferRequest":{"type":"object","properties":{"to":{"type":"string"},"amount":{"type":"number"}}},"TransferResult":{"type":"object","properties":{"transferred":{"type":"number"},"new_amount":{"type":"number"}}},"SignupRequest":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"name":{"type":"string"},"lastname":{"type":"string"},"start_amount":{"type":"number"},"middle_name":{"type":"string"},"birth_date":{"type":"string"}}},"SignupResult":{"type":"object","properties":{"client_id":{"type":"integer","format":"int64"}}},"UpdateTokenRequest":{"type":"object","properties":{"refresh_token":{"type":"string"}}},"TokensPair":{"type":"object","properties":{"refresh_token":{"type":"string"},"access_token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"InformationUpdateRequest":{"type":"object","properties":{"phones":{"type":"array","items":{"type":"string"}},"emails":{"type":"array","items":{"type":"string"}}}},"ClientDescription":{"type":"object","properties":{"login":{"type":"string"},"name":{"type":"string"},"lastname":{"type":"string"},"amount":{"type":"number"},"phones":{"type":"array","items":{"type":"string"}},"emails":{"type":"array","items":{"type":"string"}},"birthDate":{"type":"string"},"middle_name":{"type":"string"}}},"ClientSearchResponse":{"type":"object","properties":{"error":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ClientDescription"}}}}}}}