# src/main/java/com/example/testtaskforeffectivemobile/dtos/ServiceErrorMessage.java
package com.example.testtaskforeffectivemobile.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ServiceErrorMessage {
    private String error;
}