# src/main/java/com/example/testtaskforeffectivemobile/dtos/UpdateTokenRequest.java
package com.example.testtaskforeffectivemobile.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UpdateTokenRequest {
    @JsonProperty("refresh_token")
    private String refreshToken;
}